% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to SQL local host}
\usage{
db_connect(test = FALSE)
}
\arguments{
\item{test}{should connection be established with the cantabrica-test.db?}
}
\description{
Connect to SQL local host
}
