% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_create_tables}
\alias{db_create_tables}
\title{Create new tables if they don't exist}
\usage{
db_create_tables(
  con,
  tables = c("plantas", "siembras", "germinaciones", "hojas", "cosechas", "valores"),
  fill = FALSE
)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}

\item{tables}{character vector indicating the name of the tables to create ("plantas", "siembras", "germinaciones", "hojas", "cosechas", and "valores" by default)))}

\item{fill}{logical values indicating whether values should be filled in using db_fill_tables}
}
\description{
Create new tables if they don't exist
}
