% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_delete_row}
\alias{db_delete_row}
\title{Delete row from SQL table}
\usage{
db_delete_row(con, table, id)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}

\item{table}{character string indicating the table from which a row should be deleted}

\item{id}{character string indicating the ID of the row (plant) to be deleted}
}
\description{
Delete row from SQL table
}
