% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_delete_tables}
\alias{db_delete_tables}
\title{Delete tables}
\usage{
db_delete_tables(con)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}
}
\description{
Delete tables
}
