% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_fill_tables}
\alias{db_fill_tables}
\title{Add values to all tables other than 'valores'}
\usage{
db_fill_tables(con, n = 200)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}

\item{n}{number of observations to simulate (defaults to 200)}
}
\description{
Add values to all tables other than 'valores'
}
