% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_fill_values}
\alias{db_fill_values}
\title{Add values if table 'valores' is empty}
\usage{
db_fill_values(con)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}
}
\description{
Add values if table 'valores' is empty
}
