% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_get_values}
\alias{db_get_values}
\title{Get values for UI choices}
\usage{
db_get_values(con)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}
}
\description{
Get values for UI choices
}
