% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_summarise}
\alias{db_summarise}
\alias{db_summarize}
\title{Summarise all data into a table}
\usage{
db_summarise(con, data = NULL)

db_summarize(con, data = NULL)
}
\description{
Summarise all data into a table
}
