% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{diagnose_germination}
\alias{diagnose_germination}
\title{Diagnose germination model using CmdStan diagnostics}
\usage{
diagnose_germination(fit)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}
}
\description{
Diagnose germination model using CmdStan diagnostics
}
