% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{draw_germination}
\alias{draw_germination}
\title{Get posterior draws}
\usage{
draw_germination(fit)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}
}
\description{
Get posterior draws
}
