% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_germination}
\alias{fit_germination}
\title{Fit germination model}
\usage{
fit_germination(stan_code, data = NULL, lambda_prior = 1, save = NULL, ...)
}
\arguments{
\item{stan_code}{File containing the Stan code of the model}

\item{data}{Data frame with a row for each plant and at least one column with the germination days}

\item{lambda_prior}{Prior distribution of the lambda parameter in the Poisson distribution used to model germination days}

\item{save}{Path where the CmdStanFit object should be saved (as a RDS file)}

\item{...}{Arguments to be passed to the \code{$sample} operator}
}
\description{
Fit germination model
}
