% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_model}
\alias{fit_model}
\title{Fit germinacion model}
\usage{
fit_model(type, y = "y", especie = "especie", data, rds_path = NULL)
}
\arguments{
\item{type}{type of data to fit. One of "germinacion", "hojas", or "cosecha"}

\item{y}{response variable name in data}

\item{especie}{especie variable name in data}

\item{data}{data frame with a row for each plant and at least one}

\item{rds_path}{character string indicating the path in which to save the brmsfit object
column with the germination days}
}
\value{
a brmsfit object
}
\description{
Fit germinacion model
}
