% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{get_bandejas_vacias}
\alias{get_bandejas_vacias}
\title{Get empty trays}
\usage{
get_bandejas_vacias(
  con = NULL,
  data = NULL,
  instalaciones = NULL,
  bandejas = NULL
)
}
\description{
Get empty trays
}
