% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_loo}
\alias{get_loo}
\title{Leave-one-out cross-validation using}
\usage{
get_loo(fit)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}
}
\description{
Leave-one-out cross-validation using
}
