% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_post_preds}
\alias{get_post_preds}
\title{Get posterior predictions}
\usage{
get_post_preds(fit, n = 20)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}

\item{n}{Number of samples per observation}
}
\description{
Get posterior predictions
}
