% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_post_preds_checks}
\alias{get_post_preds_checks}
\title{Check posterior predictions}
\usage{
get_post_preds_checks(fit, n = 20)
}
\arguments{
\item{fit}{brmsfit object (output from \code{fit_germination})}

\item{n}{Number of samples per observation}
}
\description{
Check posterior predictions
}
