% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_prior_preds}
\alias{get_prior_preds}
\title{Get prior predictions}
\usage{
get_prior_preds(fit, n = 20)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}

\item{n}{Number of samples per observation}
}
\description{
Get prior predictions
}
