% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{loo_germination}
\alias{loo_germination}
\title{Leave-one-out cross-validation using}
\usage{
loo_germination(fit)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}
}
\description{
Leave-one-out cross-validation using
}
