% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{post_preds_germination}
\alias{post_preds_germination}
\title{Get posterior predictions}
\usage{
post_preds_germination(fit, n = 20)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}

\item{n}{Number of samples per observation}
}
\description{
Get posterior predictions
}
