% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{db_create_tables}
\alias{db_create_tables}
\title{Create new tables if they don't exist}
\usage{
db_create_tables(con)
}
\arguments{
\item{con}{connection to database, as returned by \code{db_connect}}
}
\description{
Create new tables if they don't exist
}
