% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_model}
\alias{fit_model}
\title{Fit germinacion model}
\usage{
fit_model(type = NULL, y = "y", especie = "especie", data)
}
\arguments{
\item{type}{Type of data to fit. One of "germinacion", "hojas", or "cosecha"}

\item{y}{Response variable name in data}

\item{especie}{Especie variable name in data}

\item{data}{Data frame with a row for each plant and at least one
column with the germination days}
}
\value{
A brmsfit object.
}
\description{
Fit germinacion model
}
