% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_model}
\alias{fit_model}
\title{Fit germinacion model}
\usage{
fit_model(
  type,
  data = NULL,
  lambda_prior_alpha = 15,
  lambda_prior_beta = 1,
  save = NULL,
  ...
)
}
\arguments{
\item{type}{Type of data to fit. One of "germinacion", "hojas", or "cosecha".}

\item{data}{Data frame with a row for each plant and at least one column with the germination days}

\item{save}{Path where the CmdStanFit object should be saved (as a RDS file)}

\item{...}{Arguments to be passed to the \code{$sample} operator}

\item{lambda_prior}{Prior distribution of the lambda parameter in the Poisson distribution used to model germination days}
}
\description{
Fit germinacion model
}
