% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Diagnose germination model using CmdStan diagnostics}
\usage{
get_diagnostics(fit)
}
\arguments{
\item{fit}{CmdStanFit object (output from \code{fit_germination})}
}
\description{
Diagnose germination model using CmdStan diagnostics
}
