% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_model}
\alias{plot_model}
\title{Plot MCMC posterior draws as a histogram}
\usage{
plot_model(fit)
}
\arguments{
\item{fit}{model fit, as returned by \code{fit_model}
(output from \code{draws_germination})}
}
\description{
Plot MCMC posterior draws as a histogram
}
